pop up for die button
new pop up implemented for Die button click
