+++ src/index.js
@@ -7,7 +7,7 @@ import { StartScene } from './scenes/start-scene';
 import { FirstScene } from './scenes/first-scene';
 import { MuteScene } from './scenes/mute-scene';
 import { init } from '@sentry/browser';
-import { PopUpScene } from './scenes/popup-scene';
+import { PopDieScene } from './scenes/popdiescene';
 
 /* Sentry */
 init({
@@ -23,7 +23,7 @@ const gameConfig = {
     dom: {
         createContainer: true
     },
-    scene: [ FirstScene, StartScene, JoinScene, DiceScene, PauseScene, MuteScene, PopUpScene ]
+    scene: [ FirstScene, StartScene, JoinScene, DiceScene, PauseScene, MuteScene, PopDieScene ]
 };
 
 var game = new Phaser.Game(gameConfig);
